+++ test/req/integration_test.exs
@@ -33,6 +33,7 @@ defmodule Req.IntegrationTest do
       put_range: 1,
       cache: 1,
       decompress_body: 1,
-      decode_body: 1
+      decode_body: 1,
+      handle_http_errors: 1
     ]
 end